Add examples source files for testing
The dump_all calls all the message types and dumps the information. This is
useful for testing.

The root CMakeLists.txt file was modified to allow building examples with the
rest of the code. The PARENT_SCOPE if statement was removed since it would
prevent the examples from being currently linked to the library.

A CMakeLists.txt file is added to the examples/ folder for building any source
files in that folder.