Make pre_parse_reply private static to Connection
This is so the run_command function can call the function from the header.
Making it static to conneciton.cpp makes this impossible. Declaring the function
as extern in the header makes the function accessible to everyone. This is the
best way to limit its access while allowing run_command to invoke it.