Fix pre_parse_reply failure check
If the JSON document root is an array, it will throw an exception when trying to
access a member or check for the existence of a member. So, check if the
document root is an array, before trying to check for the result key.

Also simplify the parsing. There is no reason to create an std::string before
parsing.