Implement SelectedMonitorChangeEvent
This was added in IPC Patch 1.3.0.

The Event enum is updated with the new event and the event_map is updated.
parse_selected_monitor_change_event is also added as a function.
The on_selected_monitor_change handler is added to the Connection class. The
SelectedMonitorChangeEvent struct is added.

The subscribe example was updated to subscribe to this event as well.

On an unknown event message being received, the payload will be added to the
IPCError message.