Make event socket non-blocking
handle_event becomes a blocking call and waits indefinitely for a message to
become available if it is not set as non-blocking.

The connect function is modified to take an additional argument specifying if
the socket is blocking. It also avoids a call to fcntl and sets the flag in the
socket function call.

The recv_message function takes an additional argument wait, which if true will
not return on any temporary error i.e. EAGAIN or EINTR. If wait is false, the
function will throw a NoMsgError if no bytes have been read. If any bytes have
been read, the function will try to process entire packet and will start
blocking. This is to avoid packet loss from interruptions in the middle. There
is nothing implemented at the moment to stop and resume reading/writing a
packet. The main purpose of the wait argument is to avoid the main blocking call
of read in the event that no data has been received in which case it would block
indefinitely.