Add safer disconnection handling
Add functions to connect, disconnect, and check socket connection status. Since
a disconnection can occur unexpectedly, it can be dangerous to start using
functions on a closed socket. Without these functions, once a disconnection
occurs, the class would be practically unusable.

An auto-reconnect method was considered. Although this would be convenient in
some cases, it could make it difficult for the library user to know the status
of a socket connection or if the file descriptor changed. This would especially
be unhelpful to event-driven programs which may epoll the socket file
descriptor, since auto-reconnecting would result in the file descriptor changing
without the user's knowledge. In the future, this method as an additional option
to manual control of the socket can be considered.

Special handling code was added to to deal with situations where the user may
try to write/receive messages on a disconnected socket or try to
disconnect/connect to an already disconnected/connected socket. An appropriate
exception would be thrown in this case to avoid unpredictable behaviour in
attempting to read/write to invalid sockets.

If a disconnection occurs during a write/read operation, the Connection class
will automatically close the file descriptor and mark it as disconnected, before
re-throwing the exception to the user to once again avoid unpredictable
behavior. As a result of this modifying the class, all the main public functions
had to be marked as non-const.

The main_sockfd and event_sockfd variables are no longer const and are now
private, since they may change through disconnections. The public
get_main_socket_fd and get_event_socket_fd functions can be used to retreive the
file descriptors from outside the class.

disconnect was made a utility function for closing file descriptors properly,
since its more of a generic function.

Since the socket can now be disconnected/connected by the user on-demand, the
constructor has an additional connect parameter to specify if the connection
should be made in the constructor automatically.