Add resubscribe
This function is called in connect_event_fd to resubscribe to the events
specified in subscriptions. In the initial construction of Connection, this will
result in no events being subscribed to, but on subsequent calls to
connect_event_fd after disconnections, resubscribe will resubscribe to any
events the client was subscribed to before the disconnection.

Also update subscribe.cpp example to recover from a potential disconnection.