Check for broken connection in is_*_connected()
Check for a broken connection in is_main_socket_connected and
is_event_socket_connected. If there is a broken connection, close the fd and
invalidate the socket fd by setting it to -1.

This can help safely catch broken connections without throwing exceptions and
close the file descriptors gracefully. This way, a broken connected can be
manually detected rather than detecting it by attempting a message read/write.

The is_socket_alive function attmepts to write 0 bytes to the dwm socket which
will return < 0 and set errno appropriately if the connection is broken. This
can be used to check for broken connections.

This eliminates the need for any type of ping command which would be a waste of
resources and could also confuse event replies if the connection is not broken.